Change RtpRtcp::SetRemb signature to match RtcpTransceiver::SetRemb
in particular change bitrate type to int64_t to follow style guide.

With an extra interface it will allow to add both RtpRtcp module
and RtcpTransceiver as feedback sender to PacketRouter

Bug: webrtc:8239

+++ modules/rtp_rtcp/source/rtcp_transceiver_impl.cc
@@ -110,7 +110,7 @@ void RtcpTransceiverImpl::SendCompoundPacket() {
   ReschedulePeriodicCompoundPackets();
 }
 
-void RtcpTransceiverImpl::SetRemb(int bitrate_bps,
+void RtcpTransceiverImpl::SetRemb(int64_t bitrate_bps,
                                   std::vector<uint32_t> ssrcs) {
   RTC_DCHECK_GE(bitrate_bps, 0);
   remb_.emplace();
